Extract HeaderExpectation.
